- Remove .scalafmt.config - Move the resource files for verilator into a better named scope - Removed old code - Renamed verilator tests files to be specifically for verilator   - These should ultimately be handled by some more generic backend selection mechanism - Moved utility bit twiddling code into verilator specific utils - Add and Fix OptionsAdapters for each backend in order to handle vcd writing properly   - Writing VCD files is the default for verilator, it use an explicit suppress annotation   - Not writing them is the default for Treadle   - So there is a testers2 WriteVcdAnnotation and that is mapped to the suppress or the treadle form
