--- .scalafmt.config
@@ -1,15 +1,0 @@
-style = defaultWithAlign
-
-align.openParenCallSite = false
-align.openParenDefnSite = false
-align.tokens = [{code = “->“}, {code = “<-“}, {code = “=>“, owner = “Case”}]
-continuationIndent.callSite = 2
-continuationIndent.defnSite = 2
-danglingParentheses = true
-indentOperator = spray
-maxColumn = 100
-newlines.alwaysBeforeTopLevelStatements = true
-project.excludeFilters = [“.*\\.sbt”]
-rewrite.rules = [RedundantParens, SortImports]
-spaces.inImportCurlyBraces = false
-unindentTopLevelOperators = true
