+++ src/main/scala/chisel3/tester/backends/treadle/OptionsAdapter.scala
@@ -1,0 +1,24 @@
+// See LICENSE for license details.
+
+package chisel3.tester.backends.treadle
+
+import firrtl.AnnotationSeq
+import firrtl.options.Phase
+
+/** This provides a mechanism to map between generic testers2 level annotations and
+  * any differences the backend may have in expressing their intent.
+  *
+  */
+class OptionsAdapter extends Phase {
+  /** Currently the WriteVcd annotation must be mapped to Treadle's notion of that
+    *
+    * @param annos the annotation list
+    * @return
+    */
+  override def transform(annos: AnnotationSeq): AnnotationSeq = {
+    annos.map {
+      case chisel3.tester.internal.WriteVcdAnnotation => treadle.WriteVcdAnnotation
+      case anno => anno
+    }
+  }
+}
