+++ src/main/scala/chisel3/tester/experimental/backends/verilator/Utils.scala
@@ -9,6 +9,7 @@ import chisel3.experimental._
 import chisel3.internal.InstanceId
 import chisel3.internal.firrtl.Circuit
 import chisel3.tester.backends.verilator.CommandEditor
+import treadle.utils.BitMasks
 
 import scala.collection.mutable.ArrayBuffer
 import scala.sys.process._
@@ -409,3 +410,28 @@ private[tester] object TesterProcess {
 //  def kill(p: FirrtlTerpBackend) {
 //  }
 }
+
+object Utils {
+
+  /** Converts an unsigned BigInt of width X to it's signed value
+    * Basically if msb is set convert it to a negative number
+    *
+    * @param unsigned an unsigned value that should be converted to signed
+    * @param width    the width of the target (this helps to determine if sign (MSB) is set
+    * @return         a signed BigInt
+    */
+  def unsignedBigIntToSigned(unsigned: BigInt, width: Int): BigInt = {
+    val bitMasks = BitMasks.getBitMasksBigs(width)
+
+    if (unsigned < 0) {
+      unsigned
+    } else {
+      if (bitMasks.isMsbSet(unsigned)) {
+        (unsigned & bitMasks.allBitsMask) - bitMasks.nextPowerOfTwo
+      } else {
+        unsigned & bitMasks.allBitsMask
+      }
+    }
+  }
+}
+
